Major overhaul
I use a completelly different algorithm now: first step is to enumerate all
styles in styles.xml and content.xml, noting which ones need to be converted or
not, which ones inherit the style from a parent-style, and which ones don't
have a font so need to inherit the style from the parent in the content.

Then a recursive function iterates through the content and converts the element
text and it's children tails, remembering the parent-content converter function
as it goes more deeply.